Fixing a bug in the error handling
If the stream could not be reset it was throwing an unhandled RuntimeException that was causing the rest of the error handling logic to be short-circuited.
This resulted in confusing stack traces. This fix makes sure the the inability to reset the response stream falls back to an empty ErrorDetails.