ISP-1372 fixed behavior of Response.getEntity()
Previously Response.getEntity() would not set a returnType of Response and would throw an exception when getEntity was thrown. This change fixes that behavior so that Resources that return a Response will not get an exception when getEntity() is called. getEntity() will now return an InputStream that is the body of the http response. Users can parse that input stream as they see fit.