ISP-1373 add @ResourceMetadata
This change deprecates the old constructors that required a serviceName in favor of new constructors that get the serviceName off the ResourceMetadata annotation. This annotation will appear on the ResourceInterface and hence will be provided by service writers. Additionally, service writers can use the annotation to specify modelPackages that suppress warnings that would otherwise appear when unknown package names are imported by the resource interface.