Adds support for X-OPOWER-JsonEnvelope based responses.
Bertha and scrapi both service responses and errors wrapped in
'response' and 'error' respectively. This is not compatiable with the
default opower-rest-client-builder created clients.

This commit adds support strictly for deseriailzing responses and errors
wrapped in this style.

1. The EnvelopeJacksonProvider is responsible for extracting the
response entity from the envelope in the case of a successful request.
2. The ExceptionMapperInterceptor is responsible for extracting error
details wrapped in an error envelope.